+++ packages/core/src/guidance/index.ts
@@ -1,5 +1,6 @@
 export * from './guidance.js'
 export * from './briefing.js'
+export * from './playbook-briefing.js'
 export * from './contributors.js'
 export * from './personalisation.js'
 export * from './history.js'